Ensure dep reads GOPATH from Config.Env
