+++ context.go
@@ -46,27 +46,12 @@ func (c *Ctx) SetPaths(wd string, GOPATHs ...string) error {
 // getGOPATH returns the GOPATHs from the passed environment variables.
 // If GOPATH is not defined, fallback to defaultGOPATH().
 func getGOPATHs(env []string) []string {
-	GOPATH := getEnv(env, "GOPATH")
+	GOPATH := os.Getenv("GOPATH")
 	if GOPATH == "" {
 		GOPATH = defaultGOPATH()
 	}
 
 	return filepath.SplitList(GOPATH)
-}
-
-// getEnv returns the last instance of an environment variable.
-func getEnv(env []string, key string) string {
-	for i := len(env) - 1; i >= 0; i-- {
-		v := env[i]
-		kv := strings.SplitN(v, "=", 2)
-		if kv[0] == key {
-			if len(kv) > 1 {
-				return kv[1]
-			}
-			return ""
-		}
-	}
-	return ""
 }
 
 // defaultGOPATH gets the default GOPATH that was added in 1.8
